divide delete and soft delete

+++ condition/delete.go
@@ -9,7 +9,8 @@ import (
 type Delete[T model.Model] struct {
 	OrderLimitReturning[T]
 
-	secondaryQuery *Query[T]
+	secondaryQuery       *Query[T]
+	softDeleteColumnName string
 }
 
 // Ascending specify an ascending order when updating models
@@ -45,13 +46,17 @@ func (deleteS *Delete[T]) Limit(limit int) *Delete[T] {
 //
 // warning: in mysql preloads are not allowed
 func (deleteS *Delete[T]) Returning(dest *[]T) *Delete[T] {
-	gormDB := deleteS.secondaryQuery.cqlQuery.gormDB
+	gormDB := deleteS.query.cqlQuery.gormDB
+
+	if deleteS.secondaryQuery != nil {
+		gormDB = deleteS.secondaryQuery.cqlQuery.gormDB
+	}
 
 	if len(gormDB.Statement.Selects) > 1 ||
 		len(gormDB.Statement.Preloads) > 0 {
 		deleteS.query.addError(
 			methodError(
-				preloadsInReturningNotAllowed(deleteS.secondaryQuery.cqlQuery.Dialector()),
+				preloadsInReturningNotAllowed(deleteS.query.cqlQuery.Dialector()),
 				"Returning",
 			),
 		)
@@ -69,6 +74,10 @@ func (deleteS *Delete[T]) Exec() (int64, error) {
 		return 0, deleteS.query.err
 	}
 
+	if deleteS.softDeleteColumnName != "" {
+		return deleteS.query.cqlQuery.SoftDelete(deleteS.softDeleteColumnName)
+	}
+
 	return deleteS.query.cqlQuery.Delete(
 		deleteS.secondaryQuery.cqlQuery,
 	)
@@ -82,18 +91,31 @@ func NewDelete[T model.Model](tx *gorm.DB, conditions []Condition[T]) *Delete[T]
 		err = methodError(ErrEmptyConditions, "Delete")
 	}
 
-	primaryQuery := NewQuery[T](tx)
+	model := *new(T)
+
+	var primaryQuery, secondaryQuery *Query[T]
+
+	softDeleteColumnName := model.SoftDeleteColumnName()
+
+	if softDeleteColumnName != "" {
+		// as soft delete is implemented with UPDATE, conditions can be applied directly to primary query
+		primaryQuery = NewQuery(tx, conditions...)
+	} else {
+		// for DELETE statements, a secondary query is necessary
+		primaryQuery = NewQuery[T](tx)
+		secondaryQuery = NewQuery(tx, conditions...)
+	}
+
 	if err != nil {
 		primaryQuery.err = err
 	}
-
-	secondaryQuery := NewQuery(tx, conditions...)
 
 	return &Delete[T]{
 		OrderLimitReturning: OrderLimitReturning[T]{
 			query:         primaryQuery,
 			orderByCalled: false,
 		},
-		secondaryQuery: secondaryQuery,
+		secondaryQuery:       secondaryQuery,
+		softDeleteColumnName: softDeleteColumnName,
 	}
 }
